requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
Django==5.2.1
django-browser-reload==1.18.0
django-crispy-forms==2.4
django-tailwind==4.0.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tweepy==4.15.0
types-python-dateutil==2.9.0.20250516
urllib3==2.4.0
